Vehicle/gcc-hid-main/pyproject.toml
[project]
name = "gcc-hid"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    { name = "sangaje", email = "qhals8380@gmail.com" },
]
dependencies = [
    "dbus-python>=1.4.0",
]
requires-python = "==3.10.*"
readme = "README.md"
license = { text = "MIT" }

[tool.pdm]
distribution = false

[tool.pdm.build]
package-dir = "src"
includes = ["gcc"]
